chore: update icons gulpfile.js to use ESM
